Refactor phase 3c: split the radio settings routes into a package
webapp/routes_settings_radio.py was 2,781 lines: eight module-level
helpers plus a 2,114-line register() with 26 route handlers nested
inside it. Walking the AST for names resolving to register's scope
showed every handler captures only app and route_logger (one also a
local helper, which moved with it), so the bodies move verbatim into
topic modules that each keep their own small register(app, route_logger)
-- no reindentation anywhere.

  deps.py                          129  the injectable seams
  sdr_client.py                     64  _send_sdr_command
  serialization.py                 162  receiver row -> API payload
  payload.py                       291  _parse_receiver_payload
  sync.py                          174  database <-> RadioManager
  routes_pages.py                   66  the two rendered pages
  routes_receivers.py              259  receiver CRUD
  routes_receiver_control.py       241  restart, audio-monitor wiring
  routes_devices.py                265  discovery and capabilities
  routes_presets.py                 49  tuning presets
  routes_signal.py                 355  waveform, spectrum
  routes_monitoring.py              64  dashboard status
  routes_diagnostics_status.py     306  status, SoapySDR error decoding
  routes_diagnostics_capture.py    275  IQ capture and download
  routes_diagnostics_waterfall.py  315  the waterfall view
  routes_diagnostics_analyze.py    386  analysis and auto-gain
  __init__.py                      105  register(), fanning out

webapp/routes_settings_radio.py stays as a 52-line shim. All 17 modules
are within the 400-line guidance.

The one deliberate deviation from verbatim motion is deps.py. The radio
tests inject fakes for get_redis_client, get_radio_manager,
_log_radio_event and the RADIO_CAPTURE_* constants -- names used by up
to 19 definitions each, which after the split are spread across ten
modules. A by-value import would snapshot the real object, so a stub set
in one place would be silently ignored everywhere else. Route modules
call through the module instead, giving one patch point; 51 references
were routed this way and every other line is untouched. The shim
deliberately does not re-export those names, so a stale monkeypatch
raises AttributeError rather than patching something nothing reads --
which is what made retargeting the 11 patch sites safe.

Verification.

  - 34 of 34 moved definitions are ast.dump()-identical, plus the
    every-non-blank-line-placed-exactly-once assertion (2,394 placed;
    the 54 unplaced are the licence, imports, signature and __all__).
  - URL map diffed against a worktree at the pre-split commit: 549
    rules, 0 differences. Mutation-checked by dropping a module from
    the registration tuple, which the diff caught.
  - _module_logger was logging.getLogger(__name__), which would have
    silently become webapp.radio_settings.deps; it is pinned to the
    pre-split name, and register() derives the same child logger once
    and passes it down.

Full suite green apart from a pre-existing flaky noise test
(test_8khz_with_increasing_noise seeds nothing), which passes in
isolation on both trees.

Three tooling traps are recorded in the plan's pre-split checklist: a
partition() that matched "(KR8MER)" in the copyright header instead of
the import block, a regex rewrite that cannot see scope, and ast.walk()
descending into nested functions when computing a scope's bindings.

Refs docs/development/LARGE_FILE_REFACTOR_PLAN.md
